requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
boto3==1.35.47
botocore==1.35.47
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
crispy-tailwind==1.0.3
cryptography==43.0.1
daphne==4.1.2
dj-database-url==2.3.0
Django==5.1.1
django-crispy-forms==2.3
django-environ==0.11.2
django-filter==24.3
django-paypal==2.1
django-storages==1.14.4
django-utils-six==2.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.1
zope.interface==7.0.3
